check existance of files during install (resolves #4)

+++ install.sh
@@ -12,7 +12,6 @@ if [ "$EUID" -ne 0 ]
 fi
 
 # Make sure prerequisites are installed, if not, install them
-
 if [ ! dpkg -l stlink &> /dev/null ]; then
     printf "Package stlink does not exist, please install to continue!"
     exit 1;
@@ -24,7 +23,6 @@ if [ ! dpkg -l at &> /dev/null ]; then
 fi
 
 # Make sure that the file locations exist
-
 if [ ! -d $SCRIPT_LOCATION ]; then
     printf "$SCRIPT_LOCATION does not exists, creating directory..."
     mkdir -p -m 755 $SCRIPT_LOCATION > /dev/null
@@ -35,8 +33,13 @@ if [ ! -d $UDEV_LOCATION ]; then
     mkdir -p -m 755 $UDEV_RULES_LOCATION > /dev/null
 fi
 
-# Copy files to the correct place
+# Verify that this directory has the correct files to copy
+if [ ! \( -f $UDEV_RULE -a -f $SCRIPT\)]; then
+    printf "Incorrect run location: files to copy not found."
+    exit 1;
+fi
 
+# Copy files to the correct place
 printf "Copying $SCRIPT to $SCRIPT_LOCATION/$SCRIPT..."
 cp $SCRIPT $SCRIPT_LOCATION/$SCRIPT > /dev/null
 chmod +x $SCRIPT_LOCATION/$SCRIPT
@@ -45,7 +48,6 @@ printf "Copying $UDEV_RULE to $UDEV_RULES_LOCATION/$UDEV_RULE..."
 cp $UDEV_RULE $UDEV_RULES_LOCATION/$UDEV_RULE > /dev/null
 
 # Finish up
-
 printf "Restarting udev"
 udevadm control --reload-rules && udevadm trigger
